// src/core/model/ourClient.ts
export interface OurClient {
	id?: number;
	name?: string;
	image?: string;
	clientCategoryId?: string;
	clientCategoryIdName?: string;
	isActive?: boolean;
	createDate?: Date;
	updateDate?: Date;
	deleteDate?: Date;
	createById?: number;
	updateById?: number;
	deleteById?: number;
	isDelete?: boolean;
}